refactor: improve hooks structure
