+++ assets/js/app.js
@@ -24,27 +24,14 @@ import {Socket} from "phoenix"
 import {LiveSocket} from "phoenix_live_view"
 import "../vendor/alpine.js";
 import topbar from "../vendor/topbar"
-import { QrScanner } from "./qr_reading.js";
-import { InitSorting } from "./sorting.js";
-import { StickyScroll } from "./sticky_scroll.js";
+import { QrScanner, InitSorting, StickyScroll, ScrollToTop } from "./hooks";
 
 let Hooks = {
   QrScanner: QrScanner,
   InitSorting: InitSorting,
-  StickyScroll: StickyScroll
+  StickyScroll: StickyScroll,
+  ScrollToTop: ScrollToTop
 };
-
-Hooks.ScrollToTop = {
-  mounted() {
-    this.el.addEventListener("click", e => {
-      e.preventDefault()
-      window.scrollTo({
-        top: 0,
-        behavior: 'smooth'
-      });
-    })
-  }
-}
 
 let csrfToken = document.querySelector("meta[name='csrf-token']").getAttribute("content")
 let liveSocket = new LiveSocket("/live", Socket, {
