+++ assets/js/hooks/scroll_to_top.js
@@ -1,0 +1,11 @@
+export const ScrollToTop = {
+    mounted() {
+      this.el.addEventListener("click", e => {
+        e.preventDefault()
+        window.scrollTo({
+          top: 0,
+          behavior: 'smooth'
+        });
+      })
+    }
+}
