+++ lib/atomic_web/templates/layout/_live_navbar.html.heex
@@ -191,7 +191,7 @@
               id="user-dropdown-menu"
             >
               <:wrapper>
-                <button type="button" class="-m-1.5 flex items-center p-1.5" id="user-menu-button" aria-expanded="false" aria-haspopup="true" @click="menu = ! menu" @click.away="menu = false" @keydown.escape="menu = false">
+                <button type="button" class="-m-1.5 flex items-center p-1.5" id="user-menu-button-mobile" aria-expanded="false" aria-haspopup="true" @click="menu = ! menu" @click.away="menu = false" @keydown.escape="menu = false">
                   <span class="sr-only">Open user menu</span>
                   <AtomicWeb.Components.Avatar.avatar
                     name={@current_user.name}
